requirements.txt
amqp==1.4.8
anyjson==0.3.3
beautifulsoup4==4.4.1
billiard==3.3.0.22
celery==3.1.19
decorator==4.0.4
Django==1.9
django-celery==3.1.17
django-configurations==0.8
django-extensions==1.5.9
django-filter==0.11.0
djangorestframework==3.3.1
factory-boy==2.6.0
fake-factory==0.5.3
gunicorn==19.4.1
ipython==4.0.1
ipython-genutils==0.1.0
kombu==3.0.30
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
pylibmc==1.5.0
python-dateutil==2.4.2
pytz==2015.7
requests==2.8.1
simplegeneric==0.8.1
six==1.10.0
suds==0.4
traitlets==4.0.0
wheel==0.24.0
